pyproject.toml
[project]
name = "payment-platform"
version = "0.0.1"
description = "Платформа предоставляющая удобный способ взаимодействия с партнёрским API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.11",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "pyjwt[crypto]>=2.10.1",
    "uvicorn>=0.34.0",
]
